pyproject.toml
[project]
name = "dead-escape"
version = "0.1.0"
requires-python = ">=3.9"
dependencies = ["pygame-ce"]

[project.scripts]
dead-escape = "main:main"

[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
py-modules = ["main"]

[tool.uv]
no-build-package = ["pygame-ce"]
